@@ -28,56 +28,61 @@
     "postinstall": "./node_modules/.bin/nx migrate check"
   },
   "dependencies": {
-    "@angular/animations": "8.2.14",
-    "@angular/common": "8.2.14",
-    "@angular/compiler": "8.2.14",
-    "@angular/core": "8.2.14",
-    "@angular/fire": "^5.2.3",
-    "@angular/forms": "8.2.14",
-    "@angular/platform-browser": "8.2.14",
-    "@angular/platform-browser-dynamic": "8.2.14",
-    "@angular/router": "8.2.14",
+    "@angular/animations": "9.1.7",
+    "@angular/common": "9.1.7",
+    "@angular/compiler": "9.1.7",
+    "@angular/core": "9.1.7",
+    "@angular/fire": "^6.0.0",
+    "@angular/forms": "9.1.7",
+    "@angular/platform-browser": "9.1.7",
+    "@angular/platform-browser-dynamic": "9.1.7",
+    "@angular/router": "9.1.7",
     "@firebase/app": "^0.1.10",
-    "@ngrx/effects": "8.5.1",
-    "@ngrx/entity": "8.5.1",
-    "@ngrx/router-store": "8.5.1",
-    "@ngrx/store": "8.5.1",
-    "@ngrx/store-devtools": "8.5.1",
-    "@nrwl/nx": "0.7.4",
-    "bootstrap": "^4.0.0",
-    "core-js": "^2.4.1",
-    "firebase": "^7.3.0",
-    "firebase-tools": "^7.7.0",
-    "ngx-bootstrap": "^2.0.2",
-    "rxjs": "^6.5.3",
+    "@ngrx/effects": "^9.1.2",
+    "@ngrx/entity": "9.1.2",
+    "@ngrx/router-store": "9.1.2",
+    "@ngrx/store": "9.1.2",
+    "@ngrx/store-devtools": "9.1.2",
+    "@nrwl/nx": "^7.8.7",
+    "bootstrap": "^4.5.0",
+    "core-js": "^2.6.11",
+    "firebase": "^7.14.3",
+    "firebase-tools": "^7.16.2",
+    "ngx-bootstrap": "^5.6.1",
+    "rxjs": "^6.5.5",
     "tslib": "^1.10.0",
-    "zone.js": "~0.9.1"
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.803.19",
-    "@angular-devkit/core": "8.3.19",
-    "@angular-devkit/schematics": "8.3.19",
-    "@angular/cli": "8.3.19",
-    "@angular/compiler-cli": "8.2.14",
-    "@angular/language-service": "8.2.14",
-    "@nrwl/schematics": "8.7.1",
+    "@angular-devkit/build-angular": "~0.901.6",
+    "@angular-devkit/core": "9.1.6",
+    "@angular-devkit/schematics": "9.1.6",
+    "@angular/cli": "9.1.6",
+    "@angular/compiler-cli": "9.1.7",
+    "@angular/language-service": "9.1.7",
+    "@nrwl/schematics": "8.12.11",
     "@schematics/angular": "8.3.19",
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "codelyzer": "^5.0.1",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^5.1.2",
     "jasmine-core": "~2.8.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^4.4.1",
+    "karma": "^5.0.5",
     "karma-chrome-launcher": "~2.2.0",
     "karma-cli": "~1.0.1",
     "karma-coverage-istanbul-reporter": "^1.2.1",
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
     "prettier": "1.10.2",
-    "protractor": "^5.4.2",
+    "protractor": "^7.0.0",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "3.5.3"
+    "typescript": "3.8.3",
+    "@angular-devkit/architect": "<0.900 || ^0.900.0-0 || ^9.0.0-0",
+    "firebase-tools": "^7.12.0",
+    "fuzzy": "^0.1.3",
+    "inquirer": "^6.2.2",
+    "inquirer-autocomplete-prompt": "^1.0.1"
   }
 }
